extract formatter from cell context
