+++ BackofficeBundle/Resources/public/ecmascript/OpenOrchestra/Application/Application.js
@@ -31,18 +31,17 @@ import ColorPicker            from '../Service/Form/Behavior/ColorPicker'
 import Tinymce                from '../Service/Form/Behavior/Tinymce'
 import TagCondition           from '../Service/Form/Behavior/TagCondition'
 
-import SearchFormGroupManager from '../Service/Content/SearchFormGroup/Manager'
-import DateSearchFormGroup    from '../Service/Content/SearchFormGroup/DateForm'
-import TextSearchFormGroup    from '../Service/Content/SearchFormGroup/TextForm'
-import NumberSearchFormGroup  from '../Service/Content/SearchFormGroup/NumberForm'
+import SearchFormGroupManager from '../Service/SearchFormGroup/Manager'
+import DateSearchFormGroup    from '../Service/SearchFormGroup/DateForm'
+import TextSearchFormGroup    from '../Service/SearchFormGroup/TextForm'
+import NumberSearchFormGroup  from '../Service/SearchFormGroup/NumberForm'
 
-import CellFormatterManager   from '../Service/Content/CellFormatter/Manager'
-import TextCellFormatter      from '../Service/Content/CellFormatter/TextFormatter'
-import BooleanCellFormatter   from '../Service/Content/CellFormatter/BooleanFormatter'
-import DateCellFormatter      from '../Service/Content/CellFormatter/DateFormatter'
+import CellFormatterManager   from '../Service/CellFormatter/Manager'
+import TextCellFormatter      from '../Service/CellFormatter/TextFormatter'
+import BooleanCellFormatter   from '../Service/CellFormatter/BooleanFormatter'
+import DateCellFormatter      from '../Service/CellFormatter/DateFormatter'
 
 import ApplicationError       from '../Service/Error/ApplicationError'
-import AjaxError              from '../Service/Error/AjaxError'
 import TinymceManager         from '../Service/Tinymce/TinymceManager'
 
 /**
