+++ BackofficeBundle/Resources/public/ecmascript/OpenOrchestra/Service/CellFormatter/TextFormatter.js
@@ -1,0 +1,18 @@
+import AbstractCellFormatter from './AbstractCellFormatter'
+import TextDataFormatter     from '../DataFormatter/TextFormatter'
+
+/**
+ * @class TextFormatter
+ */
+class TextFormatter extends AbstractCellFormatter
+{
+    /**
+     * @inheritdoc
+     */
+    getDataFormatter() {
+        return TextDataFormatter;
+    }
+}
+
+// unique instance of TextFormatter
+export default (new TextFormatter);
