+++ BackofficeBundle/Resources/public/ecmascript/OpenOrchestra/Service/DataFormatter/AbstractDataFormatter.js
@@ -1,0 +1,26 @@
+/**
+ * @class AbstractDataFormatter
+ */
+class AbstractDataFormatter
+{
+    /**
+     * return supported type
+     *
+     * @return string
+     */
+    getType() {
+        throw new Error('Missing getType method');
+    }
+
+    /**
+     * render the field
+     *
+     * @param   {string} value
+     * @returns {string}
+     */
+    format(value) {
+        throw new Error('Missing format method');
+    }
+}
+
+export default AbstractDataFormatter;
