+++ BackofficeBundle/Resources/public/ecmascript/OpenOrchestra/Service/SearchFormGroup/AbstractSearchFormGroup.js
