+++ BackofficeBundle/Resources/public/ecmascript/OpenOrchestra/Service/SearchFormGroup/Manager.js
